Use floats for CharacterState maxWalkSpeed and jumpSpeed
Change Mario and Mario Twin's movement speeds for each coin picked up.

+++ kindred-sprite-demo/Assets/Scripts/Character/CharacterState.cs
@@ -9,8 +9,8 @@ public class CharacterState : MonoBehaviour {
     public Vector2 position;
     public Vector2 facing;
     public Vector2 velocity;
-    public int maxWalkSpeed;
-    public int jumpSpeed;
+    public float maxWalkSpeed;
+    public float jumpSpeed;
     public int health;
     public bool isGrounded;
     public bool isWalking;
